requirements.txt
PyMuPDF==1.24.1
Pillow==10.3.0
pytesseract==0.3.10
